Revert "Revert "Warning pt.3 debruijn, include""
This reverts commit 5422223b912b6ff00e52a13383de468900f11ff4.

+++ assembler/src/debruijn/edge_index.hpp
@@ -43,9 +43,11 @@ private:
 
 public:
 
-    EdgeIndex(const Graph& g, size_t k, const std::string &workdir) :
-    GraphActionHandler<Graph>(g, "EdgeIndex"), inner_index_(k, g, workdir),
-    updater_(g, inner_index_), delete_index_(true) {
+    EdgeIndex(const Graph& g, size_t k, const std::string &workdir)
+            : GraphActionHandler<Graph>(g, "EdgeIndex"),
+              inner_index_((unsigned) k, g, workdir),
+              updater_(g, inner_index_),
+              delete_index_(true) {
     }
 
     virtual ~EdgeIndex() {
